Supports fp32, bf16, fp16 and fp8 for llm models on xpu device
