+++ python/sglang/srt/model_executor/cuda_graph_runner.py
@@ -34,7 +34,7 @@ from sglang.srt.model_executor.forward_batch_info import (
     ForwardBatch,
     ForwardMode,
 )
-from sglang.srt.utils import get_available_gpu_memory, is_hip
+from sglang.srt.utils import get_available_gpu_memory, is_hip, set_torch_compile_config
 
 _is_hip = is_hip()
 
@@ -93,20 +93,6 @@ def patch_model(
         if enable_compile:
             _to_torch(model, reverse=True, num_tokens=num_tokens)
             tp_group.ca_comm = backup_ca_comm
-
-
-def set_torch_compile_config():
-    import torch._dynamo.config
-    import torch._inductor.config
-
-    torch._inductor.config.coordinate_descent_tuning = True
-    torch._inductor.config.triton.unique_kernel_names = True
-    torch._inductor.config.fx_graph_cache = True  # Experimental feature to reduce compilation times, will be on by default in future
-
-    # FIXME: tmp workaround
-    torch._dynamo.config.accumulated_cache_size_limit = 1024
-    if hasattr(torch._dynamo.config, "cache_size_limit"):
-        torch._dynamo.config.cache_size_limit = 1024
 
 
 def get_batch_sizes_to_capture(model_runner: ModelRunner):
