+++ python/sglang/srt/utils.py
@@ -885,6 +885,20 @@ def is_triton_3():
     return triton.__version__.startswith("3.")
 
 
+def set_torch_compile_config():
+    import torch._dynamo.config
+    import torch._inductor.config
+
+    torch._inductor.config.coordinate_descent_tuning = True
+    torch._inductor.config.triton.unique_kernel_names = True
+    torch._inductor.config.fx_graph_cache = True  # Experimental feature to reduce compilation times, will be on by default in future
+
+    # FIXME: tmp workaround
+    torch._dynamo.config.accumulated_cache_size_limit = 1024
+    if hasattr(torch._dynamo.config, "cache_size_limit"):
+        torch._dynamo.config.cache_size_limit = 1024
+
+
 def maybe_torch_compile(*args, **kwargs):
     """
     torch.compile does not work for triton 2.2.0, which is needed in xlm1's jax.
